+++ sdk/src/org.graalvm.polyglot/src/org/graalvm/polyglot/impl/AbstractPolyglotImpl.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2017, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -649,6 +649,7 @@ public abstract class AbstractPolyglotImpl {
 
         public abstract <T> T lookup(Object receiver, Class<T> type);
 
+        public abstract String getWebsite(Object receiver);
     }
 
     public abstract static class AbstractLanguageDispatch extends AbstractDispatchClass {
@@ -672,6 +673,8 @@ public abstract class AbstractPolyglotImpl {
         public abstract Set<String> getMimeTypes(Object receiver);
 
         public abstract String getDefaultMimeType(Object receiver);
+
+        public abstract String getWebsite(Object receiver);
     }
 
     public abstract static class AbstractHostAccess extends AbstractDispatchClass {